fix(glm-multivariate): selaraskan output One-Way MANOVA dengan SPSS
Perbaikan numerik (Rust):
- Intercept Type III SS: ganti formula N x grandmean2 (Type I) dengan
  formula unweighted (Sigma_i ybar_i)2 / Sigma_i(1/n_i) pada
  between_subjects_effects.rs dan multivariate_tests.rs agar sesuai SPSS
  untuk desain tidak seimbang (unbalanced design)
- Standard deviation: ganti pembagi N menjadi N-1 (koreksi Bessel)
  pada calculate_variance di common.rs sehingga SD deskriptif sesuai SPSS
- Descriptive statistics: perbaiki akses data di descriptive_statistics.rs
  menggunakan merge_records sehingga nilai faktor dan DV disejajarkan per baris;
  tambah baris Total per variabel dependen
- Levene test: perbaiki akses data di levene_test.rs menggunakan merge_records
  (bug sama -- mencari nilai faktor di record DV)
- Levene df2 adjusted: implementasikan aproksimasi Welch-Satterthwaite untuk
  varian "Based on Median and with adjusted df" -- df2 kini berupa nilai
  pecahan sesuai SPSS (contoh: 2.439, 2.306)
- Bartlett test: perbaiki akses data di bartlett_test.rs menggunakan
  merge_records sehingga semua DV dapat terbaca dalam satu baris
- Box's M test: implementasikan formula dua-kasus Box (1949) untuk df2 dan
  p-value -- gunakan chi-square CDF saat c2 <= c12 agar p-value mendekati SPSS
- LeveneResult.df2: ubah tipe dari usize menjadi f64 untuk mendukung nilai pecahan

Perbaikan tampilan (TypeScript):
- Tabel Descriptive Statistics: gabungkan menjadi satu tabel (DV | level |
  Mean | Std. Deviation | N) sesuai format SPSS; urutkan DV secara alfanumerik
  agar y1 tampil sebelum y2
- Tabel Levene test: gabungkan menjadi satu tabel (DV | basis | Levene
  Statistic | df1 | df2 | Sig.) sesuai format SPSS
- Bartlett's Test of Sphericity: sembunyikan dari output GLM Multivariate
  (milik konteks Factor Analysis, bukan MANOVA); fungsi Rust tetap dihitung
  untuk potensi reuse di modul Factor Analysis mendatang
- Output renderer: perbarui filter key untuk mengenali tabel gabungan baru
  (descriptive_statistics, levene_test) sekaligus backward-compat

+++ frontend/components/Modals/Analyze/general-linear-model/multivariate/rust/src/models/result.rs
@@ -68,7 +68,7 @@ pub struct LeveneTest {
 pub struct LeveneResult {
     pub levene_statistic: f64,
     pub df1: usize,
-    pub df2: usize,
+    pub df2: f64,
     pub significance: f64,
     pub function: Option<String>,
     pub design: Option<String>,
